Fix/structr restore

+++ Units/deploy-tools/deploy-tools.sh
@@ -92,12 +92,12 @@ function deploy-tools.checkURL() {
   comment=$1
   shift
   logVerbose
-  logVerbose call: curl -k -s -o /dev/null -w "%{http_code}" "$@"
+  logVerbose call: curl --connect-timeout 0 -ksL -m 10 -o /dev/null -w "%{http_code}" "$@"
   # cUrl option -L follows redirects, e.g. if http code is 301
-  up=$(curl -k -s -L -o /dev/null -w "%{http_code}" "$@")
+  up=$(curl --connect-timeout 0 -ksL -m 10 -o /dev/null -w "%{http_code}" "$@")
   if [[ "$up" != "200" && "$up" != "302" ]]; then
     log "--: not running (returned $up): $1 - $comment"
-    curl -k -s "$@"
+    curl --connect-timeout 0 -ksL -m 10 "$@"
     return 1
   else
     log "OK: running: $1 - $comment"
@@ -179,6 +179,11 @@ function deploy-tools.checkAndCreateDataVolume() {
   fi
 
   if [[ $datavolume == *"/"* ]]; then
+    # if relative path, prepend CONFIG_DIR
+    if [[ $datavolume == .* ]]; then
+      local datavolume="${datavolume/#./$CONFIG_DIR}"
+    fi
+
     log "Volume name contains a slash, so it is a path: $datavolume"
     mkdir -p $datavolume
     chmod 777 $datavolume
@@ -277,25 +282,32 @@ function deploy-tools.recreateKeystore() {
 }
 
 function deploy-tools.checkAndRestoreDataVolume() {
-  restoresource=$1
-  datavolume=$2
-  stripcomments=$3
+  local restoresource=$1
+  local datavolume=$2
+  local stripcomponents=$3
 
   # If there is a restore source (!=none), download the file
   if [ "$restoresource" != "none" ]; then
     banner "Restore data backup"
     mkdir -p _data_restore
-    deploy-tools.downloadFile $restoresource _data_restore/data.tar.gz
+    # use shell expansion to get last part of path for $datavolume
+    deploy-tools.downloadFile $restoresource _data_restore/${datavolume##*/}.tar.gz
 
     # Move data to volume if empty
     if [[ $datavolume == *"/"* ]]; then
+      # if relative path, prepend CONFIG_DIR
+      if [[ $datavolume == .* ]]; then
+        local datavolume="${datavolume/#./$CONFIG_DIR}"
+      fi
+
       # Move data to data dir if empty
       if [ "$(ls -A $datavolume)" ]; then
         logError "Data dir is not empty: $datavolume (skip restore)"
       else
         # Extract data and strip /var/jenkins_home from the tar
         log "Extracting data into directory: $datavolume"
-        tar -xzf _data_restore/data.tar.gz -C $datavolume --strip-components=$stripcomments
+        # use shell expansion to get last part of path for $datavolume
+        tar -xzf _data_restore/${datavolume##*/}.tar.gz -C $datavolume --strip-components=$stripcomponents
       fi
     else
       files=$(docker run --rm -v $datavolume:/data alpine sh -c "ls -A /data")
@@ -304,7 +316,8 @@ function deploy-tools.checkAndRestoreDataVolume() {
       else
         # Extract data and strip /var/jenkins_home from the tar
         log "Extracting data into volume: $datavolume"
-        docker run --rm -v $datavolume:/data -v ./_data_restore:/backup alpine sh -c "tar -xzf /backup/data.tar.gz -C /data --strip-components=$stripcomments > /dev/null"
+        # use shell expansion to get last part of path for $datavolume
+        docker run --rm -v $datavolume:/data -v ./_data_restore:/backup alpine sh -c "tar -xzf /backup/${datavolume##*/}.tar.gz -C /data --strip-components=$stripcomponents > /dev/null"
         docker run --rm -v $datavolume:/data -v ./_data_restore:/backup alpine sh -c "chown -R 1000:1000 /data > /dev/null"
       fi
     fi
@@ -356,14 +369,7 @@ function deploy-tools.down() {
   CLEANUP=""
   if [ "$SCENARIO_DATA_EXTERNAL" == "false" ]; then
     CLEANUP="--volumes"
-  fi
-  docker-compose -p $SCENARIO_NAME $COMPOSE_FILE_ARGUMENTS down $CLEANUP
-  if [ "$VERBOSITY" == "-v" ]; then
-    docker ps | grep $SCENARIO_NAME
-  fi
 
-  # Remove data directory if it is a path and SCENARIO_DATA_EXTERNAL is false
-  if [[ $SCENARIO_DATA_VOLUME == *"/"* && "$SCENARIO_DATA_EXTERNAL" == "false" ]]; then
     # set separator for handling of arrays as environment variables
     IFS=','
 
@@ -371,12 +377,19 @@ function deploy-tools.down() {
     read -r -a mountpoints_array <<< "$SCENARIO_DATA_MOUNTPOINTS"
 
     for volume in "${mountpoints_array[@]}"; do
-      log "Removing data directory: $volume"
-      rm -rf $volume
+      # Remove data directory if it is a path
+      if [[ $volume == *"/"* ]]; then
+        log "Removing data directory: $volume"
+        rm -rf $volume
+      fi
     done
 
     # set separator to default value, otherwise docker-compose command will fail
     IFS=' '
+  fi
+  docker-compose -p $SCENARIO_NAME $COMPOSE_FILE_ARGUMENTS down $CLEANUP
+  if [ "$VERBOSITY" == "-v" ]; then
+    docker ps | grep $SCENARIO_NAME
   fi
 
   # Cleanup docker
